refactor(steward-list): split monolithic plugin file into directory
The 1257-line core/steward_list_plugin.rs had three sharply separated
layers (the StewardList data type, the plug-in trait, and the
DeterministicStewardList reference impl). Split into a directory with
one file per layer; each module owns its own tests.
